+++ JudoPayDotNetIntegrationTests/IntegrationTestsBase.cs
@@ -2,11 +2,10 @@
 using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Runtime.Versioning;
 using System.Text;
 using System.Threading.Tasks;
 using JudoPayDotNet;
+using JudoPayDotNet.Enums;
 using JudoPayDotNet.Http;
 using JudoPayDotNet.Models;
 using Newtonsoft.Json;
@@ -29,11 +28,15 @@ namespace JudoPayDotNetIntegrationTests
         }
 
         protected CardPaymentModel GetCardPaymentModel(
-            string yourConsumerReference = null, 
-            string cardNumber = "4976000000003436", 
-            string cv2 = "452", 
-            string postCode = "TR14 8PA", 
-            bool? recurringPayment = null, 
+            string yourConsumerReference = null,
+            string cardNumber = "4976000000003436",
+            string cv2 = "452",
+            string postCode = "TR14 8PA",
+            bool? initialRecurringPayment = null,
+            bool? recurringPayment = null,
+            string relatedReceiptId = null,
+            RecurringPaymentType? recurringPaymentType = null,
+
             string judoId = null)
         {
             if (string.IsNullOrEmpty(yourConsumerReference))
@@ -41,7 +44,7 @@ namespace JudoPayDotNetIntegrationTests
                 yourConsumerReference = Guid.NewGuid().ToString();
             }
 
-            return new CardPaymentModel
+            var ret = new CardPaymentModel
             {
                 JudoId = judoId ?? Configuration.Judoid,
                 YourConsumerReference = yourConsumerReference,
@@ -54,9 +57,25 @@ namespace JudoPayDotNetIntegrationTests
                     Line1 = "32 Edward Street",
                     PostCode = postCode,
                     Town = "Camborne"
-                },
-                RecurringPayment = recurringPayment
+                }
             };
+            if (initialRecurringPayment != null)
+            {
+                ret.InitialRecurringPayment = initialRecurringPayment;
+            }
+            if (recurringPayment != null)
+            {
+                ret.RecurringPayment = recurringPayment;
+            }
+            if (relatedReceiptId != null)
+            {
+                ret.RelatedReceiptId = relatedReceiptId;
+            }
+            if (recurringPaymentType != null)
+            {
+                ret.RecurringPaymentType = recurringPaymentType;
+            }
+            return ret;
         }
 
         protected TokenPaymentModel GetTokenPaymentModel(
@@ -302,7 +321,7 @@ namespace JudoPayDotNetIntegrationTests
                 ClientIpAddress = "127.0.0.1",
                 CompanyName = "Test",
                 Currency = "GBP",
-                ExpiryDate = DateTimeOffset.Now,
+                ExpiryDate = DateTimeOffset.Now.AddMinutes(30),
                 JudoId = Configuration.Judoid,
                 PartnerServiceFee = 10,
                 PaymentCancelUrl = "http://test.com",